+++ Core/DTOs/InventoryCounting/InventoryCountingCreationDataResponse.cs
@@ -2,14 +2,14 @@ namespace Core.DTOs.InventoryCounting;
 
 public class InventoryCountingCreationDataResponse {
     public string ItemCode { get; set; } = string.Empty;
-    public int CountedQuantity { get; set; }
-    public int SystemQuantity { get; set; }
-    public int Variance { get; set; }
+    public decimal CountedQuantity { get; set; }
+    public decimal SystemQuantity { get; set; }
+    public decimal Variance { get; set; }
     public List<InventoryCountingCreationBinResponse> CountedBins { get; set; } = new();
 }
 
 public class InventoryCountingCreationBinResponse {
     public int BinEntry { get; set; }
-    public int CountedQuantity { get; set; }
-    public int SystemQuantity { get; set; }
+    public decimal CountedQuantity { get; set; }
+    public decimal SystemQuantity { get; set; }
 }
