+++ Core/DTOs/PickList/PickingValidationResult.cs
@@ -4,7 +4,7 @@ public class PickingValidationResult {
     public bool IsValid { get; set; }
     public int? PickEntry { get; set; }
     public int ReturnValue { get; set; }
-    public int OpenQuantity { get; set; }
+    public decimal OpenQuantity { get; set; }
     public int BinOnHand { get; set; }
     public int OnHand { get; set; }
     public string? ErrorMessage { get; set; }
